replaced container with overarching container
Fixed issue #26,  took a while to figure out the issue because changes to justification of different containers would only affect one or the other rows of circles, or affect the 2 differently. Eventually realized the top had an extra container, and then added totalContainer to align both back up in a single container with the center, since previously it was having issues aligning with the center of the page including the oversized calendar page below the timer.